added authorization with JWT token (access + refresh)

+++ src/api/routes/user.js
@@ -1,48 +1,61 @@
-var express = require('express');
-var router = express.Router();
-var mongo = require('../../../connection');
+const express = require('express');
+const mongo = require('../../../connection');
+const middlewares = require('../middlewares');
+const router = express.Router();
 
-router.get('/', async function (req, res, next) {
+router.get('/', middlewares.authenticateJWT, async function (req, res, next) {
+  const { role } = req.user;
+
+  if (role !== 'admin') {
+    return res.sendStatus(403);
+  }
+
   const users = await mongo.getUsers();
-  res.send(users);
+  res.json(users);
 });
 
-router.get('/:id', async function (req, res, next) {
+router.get('/:id', middlewares.authenticateJWT, async function (req, res, next) {
   const user = await mongo.getUser(req.params.id);
   if (user) {
-    res.send(user);
+    res.json(user);
   } else {
-    res.send({});
+    res.json({});
   }
 });
 
-router.post('/', async function (req, res, next) {
+router.post('/', middlewares.authenticateJWT, async function (req, res, next) {
   const responseId = await mongo.createUser(req.body);
   if (responseId) {
-    res.send({ id: responseId, message: 'User was successfully created' });
+    res.json({ id: responseId, message: 'User was successfully created' });
   } else {
     res.status(400);
-    res.send({ message: 'Create error' });
+    res.json({ message: 'Create error' });
   }
 });
 
-router.patch('/:id', async function (req, res, next) {
+router.patch('/:id', middlewares.authenticateJWT, async function (req, res, next) {
   const responseId = await mongo.updateUser(req.params.id, req.body);
   if (responseId) {
-    res.send({ id: responseId, message: 'User was successfully updated' });
+    res.json({ id: responseId, message: 'User was successfully updated' });
   } else {
     res.status(400);
-    res.send({ message: 'Update error' });
+    res.json({ message: 'Update error' });
   }
 });
 
-router.delete('/:id', async function (req, res, next) {
+router.delete('/:id', middlewares.authenticateJWT, async function (req, res, next) {
+  const { role } = req.user;
+
+  if (role !== 'admin') {
+    return res.sendStatus(403);
+  }
+
   const responseId = await mongo.deleteUser(req.params.id);
   if (responseId) {
-    res.send({ id: responseId, message: 'User was successfully deleted' });
+    res.json({ id: responseId, message: 'User was successfully deleted' });
   } else {
     res.status(400);
-    res.send({ message: 'Delete error' });
+    res.json({ message: 'Delete error' });
   }
 });
